improve ckan itemProps + csv props
