+++ src/converters/other.ts
@@ -25,6 +25,7 @@ import {
   nullResult,
   propsToWarnings,
   itemProperties,
+  CopyProps,
 } from "./helpers";
 import { tileErrorHandlingOptions, wmsCatalogItem } from "./WmsCatalogItem";
 import { Converter } from "../convert";
@@ -374,233 +375,6 @@ export function esriMapServerCatalogGroup(
   // }
 
   return { member, messages };
-}
-
-export function ckanCatalogGroup(
-  item: CatalogMember,
-  options: ConversionOptions
-): MemberResult {
-  // See details of what's been ported https://github.com/TerriaJS/terriajs/pull/4160
-  if (!options.partial && !is.string(item.url)) {
-    return nullResult(
-      missingRequiredProp(ModelType.CkanGroup, "url", "string", item.name)
-    );
-  }
-  const propsToCopy = [
-    "url",
-    "filterQuery",
-    "groupBy",
-    "useCombinationNameWhereMultipleResources",
-    "ungroupedTitle",
-  ];
-  const unknownProps = getUnknownProps(item, [
-    ...catalogMemberProps,
-    ...catalogMemberPropsIgnore,
-    ...propsToCopy,
-    "esriMapServerResourceFormat",
-    "wmsParameters",
-  ]);
-  const member: MemberResult["member"] = {
-    type: "ckan-group",
-    name: item.name,
-  };
-  const messages = propsToWarnings(
-    ModelType.CkanGroup,
-    unknownProps,
-    item.name
-  );
-  copyProps(item, member, [...catalogMemberProps, ...propsToCopy]);
-  if (options.copyUnknownProperties) {
-    copyProps(item, member, unknownProps);
-  }
-
-  // Convert various configurations now condensed into supported resource formats
-  const supportedResourceFormats = [];
-  if (is.string(item.esriMapServerResourceFormat)) {
-    supportedResourceFormats.push({
-      id: "ArcGIS FeatureServer",
-      formatRegex: item.esriMapServerResourceFormat,
-      // definition: {
-      //   type: "esri-featureServer",
-      // },
-    });
-  }
-  if (is.plainObject(item.wmsParameters)) {
-    supportedResourceFormats.push({
-      id: "WMS",
-      definition: {
-        type: "wms",
-        parameters: item.wmsParameters,
-      },
-    });
-  }
-  member.supportedResourceFormats = supportedResourceFormats;
-  return {
-    member,
-    messages,
-  };
-}
-
-export function ckanCatalogItem(
-  item: CatalogMember,
-  options: ConversionOptions
-): MemberResult {
-  if (!options.partial && !is.string(item.url)) {
-    return nullResult(
-      missingRequiredProp(ModelType.CkanCatalogItem, "url", "string", item.name)
-    );
-  }
-  const propsToCopy = ["url", "datasetId", "resourceId"];
-
-  const unknownProps = getUnknownProps(item, [
-    ...catalogMemberProps,
-    ...catalogMemberPropsIgnore,
-    ...propsToCopy,
-    "allowAnyResourceIfResourceIdNotFound",
-    "esriFeatureServerResourceFormat",
-    "wmsResourceFormat",
-    "wfsResourceFormat",
-    "kmlResourceFormat",
-    "csvResourceFormat",
-    "esriMapServerResourceFormat",
-    "esriFeatureServerResourceFormat",
-    "geoJsonResourceFormat",
-    "czmlResourceFormat",
-    "itemProperties",
-  ]);
-  const member: MemberResult["member"] = {
-    type: "ckan-item",
-    name: item.name,
-  };
-  const messages = propsToWarnings(
-    ModelType.CkanCatalogItem,
-    unknownProps,
-    item.name
-  );
-  copyProps(item, member, [...catalogMemberProps, ...propsToCopy]);
-  if (options.copyUnknownProperties) {
-    copyProps(item, member, unknownProps);
-  }
-
-  const itemPropertiesHelper = (converter: Converter) => {
-    if (isPlainObject(item.itemProperties)) {
-      const itemPropertiesResult = itemProperties(item, converter);
-      if (itemPropertiesResult.result)
-        member.itemProperties = itemPropertiesResult.result;
-      messages.push(...itemPropertiesResult.messages);
-    }
-  };
-
-  // Convert various configurations now condensed into supported resource formats
-  const supportedResourceFormats = [];
-  if (is.string(item.esriFeatureServerResourceFormat)) {
-    supportedResourceFormats.push({
-      id: "ArcGIS FeatureServer",
-      formatRegex: item.esriFeatureServerResourceFormat,
-      definition: {
-        type: "esri-featureServer",
-      },
-    });
-    itemPropertiesHelper(esriFeatureServerCatalogItem);
-  } else if (is.string(item.wmsResourceFormat)) {
-    supportedResourceFormats.push({
-      id: "WMS",
-      formatRegex: item.wmsResourceFormat,
-      definition: {
-        type: "wms",
-      },
-    });
-    itemPropertiesHelper(wmsCatalogItem);
-  } else if (is.string(item.wfsResourceFormat)) {
-    supportedResourceFormats.push({
-      id: "WFS",
-      formatRegex: item.wfsResourceFormat,
-      definition: {
-        type: "wfs",
-      },
-    });
-    if (isPlainObject(item.itemProperties)) {
-      messages.push({
-        message: `WFS itemProperties is not supported`,
-        path: [item.name, "itemProperties (WFS)"],
-        severity: Severity.Warning,
-        details: unknownPropOpaque.toOpaque({
-          modelType: ModelType.CkanCatalogItem,
-          property: "itemProperties (WFS)",
-        }),
-      });
-    }
-  } else if (is.string(item.kmlResourceFormat)) {
-    supportedResourceFormats.push({
-      id: "Kml",
-      formatRegex: item.kmlResourceFormat,
-      definition: {
-        type: "kml",
-      },
-    });
-    itemPropertiesHelper(kmlCatalogItem);
-  } else if (is.string(item.csvResourceFormat)) {
-    supportedResourceFormats.push({
-      id: "CSV",
-      formatRegex: item.csvResourceFormat,
-      definition: {
-        type: "csv",
-      },
-    });
-    itemPropertiesHelper(csvCatalogItem);
-  } else if (is.string(item.esriMapServerResourceFormat)) {
-    supportedResourceFormats.push({
-      id: "ArcGIS MapServer",
-      formatRegex: item.esriMapServerResourceFormat,
-      definition: {
-        type: "esri-mapServer",
-      },
-    });
-    itemPropertiesHelper(esriMapServerCatalogItem);
-  } else if (is.string(item.esriFeatureServerResourceFormat)) {
-    supportedResourceFormats.push({
-      id: "ArcGIS FeatureServer",
-      formatRegex: item.esriFeatureServerResourceFormat,
-      definition: {
-        type: "esri-featureServer",
-      },
-    });
-    itemPropertiesHelper(esriFeatureServerCatalogItem);
-  } else if (is.string(item.geoJsonResourceFormat)) {
-    supportedResourceFormats.push({
-      id: "GeoJson",
-      formatRegex: item.geoJsonResourceFormat,
-      definition: {
-        type: "geojson",
-      },
-    });
-    itemPropertiesHelper(geoJsonCatalogItem);
-  } else if (is.string(item.czmlResourceFormat)) {
-    supportedResourceFormats.push({
-      id: "Czml",
-      formatRegex: item.czmlResourceFormat,
-      definition: {
-        type: "czml",
-      },
-    });
-    if (isPlainObject(item.itemProperties)) {
-      messages.push({
-        message: `Czml itemProperties is not supported`,
-        path: [item.name, "itemProperties (Czml)"],
-        severity: Severity.Warning,
-        details: unknownPropOpaque.toOpaque({
-          modelType: ModelType.CkanCatalogItem,
-          property: "itemProperties (Czml)",
-        }),
-      });
-    }
-  }
-
-  member.supportedResourceFormats = supportedResourceFormats;
-  return {
-    member,
-    messages,
-  };
 }
 
 export function wpsCatalogItem(
